Fixed undefined shift behaviour
BOOL must be cast to UINT32 before shifting to avoid undefined
behaviour

+++ libfreerdp/core/capabilities.c
@@ -1806,7 +1806,7 @@ static void rdp_write_bitmap_cache_cell_info(wStream* s, BITMAP_CACHE_V2_CELL_IN
 	 * numEntries is in the first 31 bits, while the last bit (k)
 	 * is used to indicate a persistent bitmap cache.
 	 */
-	info = (cellInfo->numEntries | (cellInfo->persistent << 31));
+	info = (cellInfo->numEntries | (((UINT32)cellInfo->persistent) << 31));
 	Stream_Write_UINT32(s, info);
 }
